backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
groq>=0.9.0
yfinance>=0.2.40
pandas>=2.0.0
pandas-ta>=0.3.14b
python-dotenv>=1.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0
apscheduler>=3.10.0
